refactor: move the validation validaotrs to their own utils file
Refs: #SFB-3
